fix(streams_item): update subtitles size and offset types

+++ src/types/streams/streams_item.rs
@@ -34,10 +34,10 @@ pub struct StreamItemState {
     pub subtitle_delay: Option<i64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     /// Subtitles size, platform dependent units
-    pub subtitle_size: Option<u64>,
+    pub subtitle_size: Option<f32>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     /// Vertical offset of the subtitles, platform dependent units
-    pub subtitle_offset: Option<u64>,
+    pub subtitle_offset: Option<f32>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub audio_track: Option<AudioTrack>,
     /// In milliseconds
